@@ -1,67 +1,54 @@
 package com.abhishek.germanPocketDictionary.activity;
 
-import android.annotation.SuppressLint;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.net.Uri;
-import android.os.AsyncTask;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.Handler;
-import android.util.Log;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.CheckBox;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.StringRes;
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.preference.PreferenceManager;
-import androidx.viewpager.widget.ViewPager;
-
-import com.abhishek.germanPocketDictionary.R;
-import com.abhishek.germanPocketDictionary.adapter.CategoryPagerAdapter;
-import com.abhishek.germanPocketDictionary.firebase.FirebaseHandler;
-import com.abhishek.germanPocketDictionary.model.Word;
-import com.abhishek.germanPocketDictionary.utilities.ConnectionUtils;
-import com.abhishek.germanPocketDictionary.utilities.Constants;
-import com.abhishek.germanPocketDictionary.utilities.SharedPreferenceManager;
-import com.abhishek.germanPocketDictionary.utilities.Utils;
-import com.google.android.material.navigation.NavigationView;
-import com.google.android.material.tabs.TabLayout;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.ValueEventListener;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
 import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_COLORS;
 import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_NOUNS;
 import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_NUMBERS;
 import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_OPPOSITE;
 import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_QUESTIONS;
 import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_VERBS;
-import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.PREF_AGREEMENT_KEY;
 import static com.abhishek.germanPocketDictionary.utilities.Utils.getFragmentLocationFromCategory;
 
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+import android.os.Handler;
+import android.util.Log;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+import android.widget.Toast;
 
-public class MainActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.viewpager.widget.ViewPager;
+
+import com.abhishek.germanPocketDictionary.R;
+import com.abhishek.germanPocketDictionary.activity.feedback.ui.FeedBackActivity;
+import com.abhishek.germanPocketDictionary.adapter.CategoryPagerAdapter;
+import com.abhishek.germanPocketDictionary.data.WordsRepository;
+import com.abhishek.germanPocketDictionary.model.Word;
+import com.abhishek.germanPocketDictionary.utilities.Constants;
+import com.abhishek.germanPocketDictionary.utilities.Utils;
+import com.google.android.material.navigation.NavigationView;
+import com.google.android.material.tabs.TabLayout;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+import javax.inject.Inject;
+
+import dagger.hilt.android.AndroidEntryPoint;
+
+@AndroidEntryPoint
+public class MainActivity extends AppCompatActivity {
 
     public List<Word> allWordsList;
 
@@ -78,127 +65,33 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
 
     private boolean MENU_ITEM_HIDE = true;
 
-    private AlertDialog alertDialog;
-
     private ViewPager viewPager;
 
-    private Bundle savedInstanceState;
-
-    private DatabaseReference wordsReference;
-    private ValueEventListener wordsValueEventListener;
-
+    @Inject
+    WordsRepository wordsRepository;
+    private HomeViewModel homeViewModel;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         // Set the content of the activity to use the activity_main.xml layout file
         setContentView(R.layout.activity_main);
-        this.savedInstanceState = savedInstanceState;
-
 
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-        setupAgreementAlertDialog();
-    }
+        setupNavDrawer();
 
-    private void setupAgreementAlertDialog() {
-        String agreement = null;
-        try {
-            agreement = fetchAgreementDetails();
-        } catch (IOException e) {
-            agreementLoadingFailed();
-        }
+        allWordsList = new ArrayList<>();
 
-        final SharedPreferences sharedPreferences =
-                PreferenceManager.getDefaultSharedPreferences(this);
+        HomeViewModel.HomeViewModelFactory homeViewModelFactory = new HomeViewModel.HomeViewModelFactory(wordsRepository);
+        homeViewModel = ViewModelProviders.of(this, homeViewModelFactory).get(HomeViewModel.class);
+        homeViewModel.getWords();
 
-        boolean agreed = sharedPreferences.getBoolean(PREF_AGREEMENT_KEY, false);
-        if (sharedPreferences.contains("agreed")) {
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.remove("agreed");
-            editor.apply();
-        }
-
-        if (agreement != null) {
-            @SuppressLint("InflateParams") View alertDialogLayout = getLayoutInflater().inflate(R.layout.agreement_alertbox_layout, null);
-            TextView agreementTextView = alertDialogLayout.findViewById(R.id.agreement_details);
-            agreementTextView.setText(agreement);
-            if (!agreed) {
-                AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                builder.setTitle(R.string.agreement_title)
-                        .setView(alertDialogLayout)
-                        .setPositiveButton("I Agree", (dialog, which) -> {
-                            SharedPreferences.Editor editor = sharedPreferences.edit();
-                            editor.putBoolean(PREF_AGREEMENT_KEY, true);
-                            editor.apply();
-                            setupNavDrawer();
-                            setupInterface();
-                        })
-                        .setNegativeButton("I Disagree", (dialog, which) -> {
-                            Toast.makeText(this, "Application is Closing", Toast.LENGTH_SHORT).show();
-                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                                finishAndRemoveTask();
-                            } else
-                                finishAffinity();
-
-                            SharedPreferences.Editor editor = sharedPreferences.edit();
-                            editor.putBoolean(PREF_AGREEMENT_KEY, false);
-                            editor.apply();
-                        })
-                        .setCancelable(false);
-
-                alertDialog = builder.create();
-                alertDialog.show();
-
-                alertDialog.getButton(AlertDialog.BUTTON_POSITIVE).setEnabled(false);
-
-                CheckBox checkBox = alertDialogLayout.findViewById(R.id.agreement_checkbox);
-                checkBox.setOnCheckedChangeListener((buttonView, isChecked) -> {
-                    if (isChecked)
-                        updateAgreementStatus(true);
-                    else
-                        updateAgreementStatus(false);
-                });
-            } else {
-                setupNavDrawer();
-                setupInterface();
-            }
-        } else {
-            agreementLoadingFailed();
-        }
-    }
-
-    private String fetchAgreementDetails() throws IOException {
-
-        StringBuilder agreement = new StringBuilder();
-        try {
-            InputStream inputStream = getResources().openRawResource(
-                    getResources().getIdentifier("com.abhishek.germanPocketDictionary:raw/terms_of_use_warranties_and_release_agreement", null, null));
-
-            //noinspection ConstantConditions
-            if (inputStream != null) {
-                InputStreamReader inputStreamReader = new InputStreamReader(inputStream, StandardCharsets.UTF_8);
-                BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
-                String line = bufferedReader.readLine();
-                while (line != null) {
-                    agreement.append("\n");
-                    agreement.append(line);
-                    line = bufferedReader.readLine();
-                }
-            }
-            return agreement.toString();
-        } catch (Exception e) {
-            throw new IOException();
-        }
-    }
-
-    private void agreementLoadingFailed() {
-        Toast.makeText(this, "Agreement failed to load, contact the developer", Toast.LENGTH_LONG).show();
-        hideProgressBarAndShowTextView(R.string.agreement_failed_to_load);
-    }
-
-    public void updateAgreementStatus(Boolean status) {
-        alertDialog.getButton(AlertDialog.BUTTON_POSITIVE).setEnabled(status);
+        homeViewModel.getAllWords().observe(this, words -> {
+            allWordsList.clear();
+            allWordsList.addAll(words);
+            updateUI();
+        });
     }
 
     private void setupNavDrawer() {
@@ -244,71 +137,6 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
             mDrawerLayout.closeDrawer(GravityCompat.START);
             return false;
         });
-    }
-
-    public void setupInterface() {
-        allWordsList = new ArrayList<>();
-
-        attachValueEventListener();
-
-        SharedPreferenceManager.getInstance(this).registerOnSharedPreferenceChangeListener(this);
-        if (savedInstanceState != null && savedInstanceState.containsKey(Constants.TABLES.ALL_WORDS)) {
-            Log.d(LOG_TAG, "fetching json from savedInstanceBundle..");
-            String json = savedInstanceState.getString(Constants.TABLES.ALL_WORDS);
-            if (json != null && !json.equals("")) {
-                allWordsList.clear();
-                allWordsList.addAll(Utils.getWordListFromJson(json));
-                Log.d(LOG_TAG, "successfully fetched data from savedInstanceBundle");
-                updateUI();
-            } else
-                fetchWordsFromPreference();
-        } else
-            fetchWordsFromPreference();
-
-    }
-
-    private void fetchWordsFromPreference() {
-        showProgressBarAndLoadingTextView();
-        Log.d(LOG_TAG, "fetching words from the pref...");
-        SharedPreferenceManager prefManager = SharedPreferenceManager.getInstance(this);
-        if (prefManager.contains(Constants.TABLES.ALL_WORDS)) {
-            List<Word> words = prefManager.getListFromPreference(Constants.TABLES.ALL_WORDS);
-            if (words == null || words.isEmpty()) {
-                Log.d(LOG_TAG, "fetching words from db - words in pref is empty/null");
-                fetchWordsFromDb();
-            } else {
-                Log.d(LOG_TAG, "Successfully fetched words from the pref.");
-                allWordsList.clear();
-                allWordsList.addAll(words);
-                updateUI();
-            }
-        } else {
-            Log.d(LOG_TAG, "words not found in shared preferences");
-            fetchWordsFromDb();
-        }
-    }
-
-    private void fetchWordsFromDb() {
-        if (allWordsList == null)
-            allWordsList = new ArrayList<>();
-
-        Log.d(LOG_TAG, "fetching words from the firebase...");
-        ConnectionUtils connectionUtils = new ConnectionUtils(this);
-
-        if (!connectionUtils.hasInternetAccess())
-            noInternetConnection();
-
-        if (wordsReference == null)
-            wordsReference = FirebaseHandler.getInstance().getWordsReference();
-
-        wordsReference.addListenerForSingleValueEvent(getWordsValueEventListener());
-    }
-
-    private void attachValueEventListener() {
-        if (wordsReference == null)
-            wordsReference = FirebaseHandler.getInstance().getWordsReference();
-
-        wordsReference.addValueEventListener(getWordsValueEventListener());
     }
 
     private void updateUI() {
@@ -384,40 +212,6 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
         progressBar.setVisibility(View.VISIBLE);
     }
 
-    private ValueEventListener getWordsValueEventListener() {
-        if (wordsValueEventListener == null) {
-            wordsValueEventListener = new ValueEventListener() {
-                @Override
-                public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                    allWordsList.clear();
-                    for (DataSnapshot wordsSnapshot : dataSnapshot.getChildren()) {
-                        Word word = wordsSnapshot.getValue(Word.class);
-                        if (word != null) {
-                            allWordsList.add(word);
-                        }
-                    }
-
-                    if (allWordsList != null && !allWordsList.isEmpty()) {
-                        Log.d(LOG_TAG, "words successfully fetched from the db");
-                        Log.d(LOG_TAG, "updating preferences....");
-                        SharedPreferenceManager prefManager = SharedPreferenceManager
-                                .getInstance(MainActivity.this);
-                        prefManager.setList(Constants.TABLES.ALL_WORDS, allWordsList);
-                        hideProgressbarAndLoadingTextView();
-                    }
-                }
-
-                @Override
-                public void onCancelled(@NonNull DatabaseError databaseError) {
-                    Toast.makeText(MainActivity.this, "Failed to load data, please try again!", Toast.LENGTH_SHORT).show();
-                    Log.d(LOG_TAG, "Firebase onCancelled - Failed to load data, please try again!");
-                    hideProgressBarAndShowTextView(R.string.error_loading_data);
-                }
-            };
-        }
-        return wordsValueEventListener;
-    }
-
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.menu, menu);
@@ -481,13 +275,6 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
             super.onBackPressed();
     }
 
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String s) {
-        Log.d(LOG_TAG, "preference updated");
-        fetchWordsFromPreference();
-    }
-
     @Override
     protected void onSaveInstanceState(@NonNull Bundle outState) {
         Log.d(LOG_TAG, "onSaveInstanceState called");
@@ -497,14 +284,5 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
             outState.putString(Constants.TABLES.ALL_WORDS, json);
         }
         super.onSaveInstanceState(outState);
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        SharedPreferenceManager.getInstance(this).unregisterOnSharedPreferenceChangeListener(this);
-        if (wordsReference != null && wordsValueEventListener != null) {
-            wordsReference.removeEventListener(wordsValueEventListener);
-        }
     }
 }
