@@ -1,11 +1,12 @@
 package com.abhishek.germanPocketDictionary.adapter;
 
+import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_NOUNS;
+import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_OPPOSITE;
+import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_VERBS;
+import static com.abhishek.germanPocketDictionary.utilities.Utils.getWordsUsingCategory;
+
 import android.app.Activity;
 import android.content.Context;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
 import android.content.SharedPreferences;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,20 +14,18 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.abhishek.germanPocketDictionary.R;
 import com.abhishek.germanPocketDictionary.activity.SearchResultActivity;
+import com.abhishek.germanPocketDictionary.fragments.WordsFragment;
 import com.abhishek.germanPocketDictionary.interfaces.OnWordClickListener;
 import com.abhishek.germanPocketDictionary.model.Word;
-import com.abhishek.germanPocketDictionary.fragments.WordsFragment;
-import com.abhishek.germanPocketDictionary.R;
-
-import java.util.List;
-
-import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_NOUNS;
-import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_OPPOSITE;
-import static com.abhishek.germanPocketDictionary.utilities.Constants.API_KEYS.CATEGORY_VERBS;
-import static com.abhishek.germanPocketDictionary.utilities.Utils.getWordsUsingCategory;
-
 import com.abhishek.germanPocketDictionary.utilities.Constants;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author Abhishek Saxena
@@ -45,15 +44,6 @@ public class WordAdapter extends RecyclerView.Adapter<WordAdapter.ViewHolder> im
     private List<Word> words;
     private OnWordClickListener onWordClickListener;
 
-    public WordAdapter(WordsFragment fragment, Context context, String fragmentType,
-                       OnWordClickListener onWordClickListener) {
-        this.context = context;
-        this.words = getWordsUsingCategory(context, fragmentType);
-        this.fragment = fragment;
-        mFragmentType = fragmentType;
-        this.onWordClickListener = onWordClickListener;
-    }
-
     public WordAdapter(Activity context, List<Word> words,
                        OnWordClickListener onWordClickListener) {
         mFragmentType = "search_type";
@@ -63,10 +53,24 @@ public class WordAdapter extends RecyclerView.Adapter<WordAdapter.ViewHolder> im
         this.onWordClickListener = onWordClickListener;
     }
 
+    public WordAdapter(WordsFragment fragment, Context context, String fragmentType,
+                       OnWordClickListener onWordClickListener) {
+        this.context = context;
+        this.words = new ArrayList<>();
+        this.fragment = fragment;
+        mFragmentType = fragmentType;
+        this.onWordClickListener = onWordClickListener;
+    }
+
+    public void submitList(List<Word> words) {
+        this.words = words;
+        notifyDataSetChanged();
+    }
+
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(context).inflate(mFragmentType.equals(Constants.API_KEYS.CATEGORY_OPPOSITE) ? R.layout.opposite_words : R.layout.list_item, parent, false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(mFragmentType.equals(Constants.API_KEYS.CATEGORY_OPPOSITE) ? R.layout.opposite_words : R.layout.list_item, parent, false);
         return new ViewHolder(view, isActivity, onWordClickListener);
     }
 
