@@ -1,13 +1,10 @@
 package com.abhishek.germanPocketDictionary.fragments;
 
 
-import android.content.SharedPreferences;
+import static android.content.ContentValues.TAG;
+
+import android.content.Context;
 import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.core.content.ContextCompat;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 import android.transition.ChangeBounds;
 import android.transition.TransitionManager;
 import android.util.Log;
@@ -18,28 +15,43 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.abhishek.germanPocketDictionary.R;
 import com.abhishek.germanPocketDictionary.activity.MainActivity;
 import com.abhishek.germanPocketDictionary.adapter.WordAdapter;
-import com.abhishek.germanPocketDictionary.R;
+import com.abhishek.germanPocketDictionary.data.WordsRepository;
 import com.abhishek.germanPocketDictionary.interfaces.OnWordClickListener;
 import com.abhishek.germanPocketDictionary.utilities.Constants;
 
-import java.util.ArrayList;
+import javax.inject.Inject;
 
-import static android.content.ContentValues.TAG;
+import dagger.hilt.android.AndroidEntryPoint;
 
 /**
  * Created by Abhishek Saxena on 12/15/2017.
  */
 
+@AndroidEntryPoint
 public class WordsFragment extends Fragment implements OnWordClickListener {
 
     public int selectedNoun;
     public int selectedVerb;
 
-    private RecyclerView.Adapter mAdapter;
+    private WordsViewModel wordsViewModel;
+
+    private WordAdapter mAdapter;
 
     private String fragmentType;
+
+    @Inject
+    WordsRepository wordsRepository;
 
     public static WordsFragment newInstance(String type) {
         WordsFragment wordFragment = new WordsFragment();
@@ -62,37 +74,14 @@ public class WordsFragment extends Fragment implements OnWordClickListener {
         if (getArguments() != null && activityReference != null) {
             fragmentType = getArguments().getString(Constants.API_KEYS.FRAGMENT_TYPE);
             if (fragmentType != null) {
-                switch (fragmentType) {
-                    case Constants.TABLES.ALL_WORDS:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.TABLES.ALL_WORDS, this);
-                        break;
 
-                    case Constants.API_KEYS.CATEGORY_NOUNS:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.API_KEYS.CATEGORY_NOUNS, this);
-                        selectedNoun = -1;
-                        break;
+                mAdapter = createAdapter(this, getContext(), fragmentType, this);
 
-                    case Constants.API_KEYS.CATEGORY_VERBS:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.API_KEYS.CATEGORY_VERBS, this);
-                        selectedVerb = -1;
-                        break;
+                if (fragmentType.equals(Constants.API_KEYS.CATEGORY_NOUNS))
+                    selectedNoun = -1;
 
-                    case Constants.API_KEYS.CATEGORY_NUMBERS:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.API_KEYS.CATEGORY_NUMBERS, this);
-                        break;
-
-                    case Constants.API_KEYS.CATEGORY_COLORS:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.API_KEYS.CATEGORY_COLORS, this);
-                        break;
-
-                    case Constants.API_KEYS.CATEGORY_QUESTIONS:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.API_KEYS.CATEGORY_QUESTIONS, this);
-                        break;
-
-                    case Constants.API_KEYS.CATEGORY_OPPOSITE:
-                        mAdapter = new WordAdapter(this, getContext(), Constants.API_KEYS.CATEGORY_OPPOSITE, this);
-                        break;
-                }
+                if (fragmentType.equals(Constants.API_KEYS.CATEGORY_VERBS))
+                    selectedVerb = -1;
             }
         }
         LinearLayoutManager manager = new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL, false);
@@ -111,43 +100,65 @@ public class WordsFragment extends Fragment implements OnWordClickListener {
     }
 
     @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        WordsViewModel.Factory wordsViewModelFactory = new WordsViewModel.Factory(wordsRepository, fragmentType);
+        wordsViewModel = ViewModelProviders.of(this, wordsViewModelFactory).get(WordsViewModel.class);
+
+        wordsViewModel.fetchWordsByCategory();
+        observeToWords();
+    }
+
+    private void observeToWords() {
+        wordsViewModel.getWords().observe(getViewLifecycleOwner(), words -> mAdapter.submitList(words));
+    }
+
+    private WordAdapter createAdapter(WordsFragment fragment,
+                                      Context context,
+                                      String category,
+                                      OnWordClickListener onWordClickListener) {
+        return new WordAdapter(fragment, context, category, onWordClickListener);
+    }
+
+    @Override
     public void onWordClickListener(int position, View view) {
         Log.d(TAG, "onWordClickListener: clicked, position : " + position);
 
         if (getArguments() != null) {
             fragmentType = getArguments().getString(Constants.API_KEYS.FRAGMENT_TYPE);
 
-            if (fragmentType != null) {
-                switch (fragmentType) {
-                    case Constants.API_KEYS.CATEGORY_NOUNS:
-                        if (getActivity() != null) {
-                            //ViewGroup mainView = rootView.findViewById(R.id.main_view);
-                            ImageView arrowImageView = view.findViewById(R.id.arrow);
-                            TransitionManager.beginDelayedTransition((ViewGroup) view, new ChangeBounds());
-                            if (selectedNoun != position) {
-                                view.findViewById(R.id.expandable_view).setVisibility(View.VISIBLE);
+            if (fragmentType == null) return;
 
-                                arrowImageView.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_keyboard_arrow_up_black_18dp));
-                                selectedNoun = position;
-                            } else {
-                                arrowImageView.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_keyboard_arrow_up_black_18dp));
-                                view.findViewById(R.id.expandable_view).setVisibility(View.GONE);
-                                selectedNoun = -1;
-                            }
-                            mAdapter.notifyDataSetChanged();
-                        }
-                        break;
-                    case Constants.API_KEYS.CATEGORY_VERBS:
-                        if (selectedVerb != position) {
+            switch (fragmentType) {
+                case Constants.API_KEYS.CATEGORY_NOUNS:
+                    if (getActivity() != null) {
+                        //ViewGroup mainView = rootView.findViewById(R.id.main_view);
+                        ImageView arrowImageView = view.findViewById(R.id.arrow);
+                        TransitionManager.beginDelayedTransition((ViewGroup) view, new ChangeBounds());
+                        if (selectedNoun != position) {
                             view.findViewById(R.id.expandable_view).setVisibility(View.VISIBLE);
-                            selectedVerb = position;
+
+                            arrowImageView.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_keyboard_arrow_up_black_18dp));
+                            selectedNoun = position;
                         } else {
+                            arrowImageView.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_keyboard_arrow_up_black_18dp));
                             view.findViewById(R.id.expandable_view).setVisibility(View.GONE);
-                            selectedVerb = -1;
+                            selectedNoun = -1;
                         }
                         mAdapter.notifyDataSetChanged();
-                        break;
-                }
+                    }
+                    break;
+                case Constants.API_KEYS.CATEGORY_VERBS:
+                    if (selectedVerb != position) {
+                        view.findViewById(R.id.expandable_view).setVisibility(View.VISIBLE);
+                        selectedVerb = position;
+                    } else {
+                        view.findViewById(R.id.expandable_view).setVisibility(View.GONE);
+                        selectedVerb = -1;
+                    }
+                    mAdapter.notifyDataSetChanged();
+                    break;
             }
         }
     }
